Worked on databases: sorted some errors and added packages.

+++ src/main/java/com/csc207/routop/Week.java
@@ -14,7 +14,7 @@ import org.springframework.boot.autoconfigure.domain.EntityScan;
 public class Week {
     @Id
     @GeneratedValue(strategy=GenerationType.AUTO)
-    private Long Id;
+    private int Id;
     public Day[] days;
